Finish removing tokio
It had been a long while since tokio was actually required, but since
the crate was using async-lock, it was still not compatible with
non-async code. After reviewing the code again, I realized all of the
locks are short lived, and never run any user code within the locked
spans. Thus, it's safe to just use parking_lot.

After removing async from as many methods as possible, this crate is now
fully runtime agnostic and can be used in both non-async and async
contexts.